# openapi.yaml
openapi: 3.1.0
info:
  title: Event Hunter API
  version: 0.1.0
  summary: 事件驅動架構治理與業務事故鑑識 API 草案
  description: |
    Event Hunter 是建立在 Kafka、ClickHouse 與 OpenTelemetry 資料之上的
    業務級事件鑑識平台。本文件保留完整的未來 API 草案；MVP 的 Event Hunter
    使用者介面聚焦 Business Timeline、Investigation、固定 Domain Pattern 與
    Evidence Bundle，另包含 Grafana 業務告警接入、事件品質聚合與可展示的
    Outbox-to-Timeline 管線等 supporting capabilities。
    Event Catalog、Topic Registry、Event Hunter 自建 Runtime Quality UI 與 Replay
    是延後功能，先由 Schema Registry／Kafka tooling／Grafana OSS 等既有元件處理。
    Grafana OSS 負責 Logs／Metrics／Traces、Dashboard、Explore、Correlations 與
    Alerting；Event Hunter 只接收需要業務調查的 Grafana Alert Webhook，並保存參照。

    ### 重要邊界

    - 第一版 Pattern Engine 只做固定規則的唯讀調查，不直接修改正式業務資料。
    - `forensics_events` 等事件資料位於 ClickHouse；案件、權限與設定位於 PostgreSQL。
    - Temporal 是預設關閉的選用能力，只負責低頻、可重試、可能等待人工核准的案件流程；MVP 核心功能不依賴 Temporal。
    - Kafka Consumer 的事件重試由 Consumer Framework／Retry Topic／DLQ 處理；Event Hunter 只有在來源提供 attempt 與錯誤資訊時才呈現重試摘要。
    - Production Redrive（把歷史事件重新送回正式事件流）不在本 API 草案範圍內。
  contact:
    name: Event Hunter Team
  license:
    name: Proprietary / Internal Draft
servers:
  - url: http://localhost:28333
    description: 本機開發環境
  - url: https://event-hunter.example.com
    description: 預留的部署環境（尚未實作）
...
paths:
...
  /api/v1/events:
    x-mvp-status: deferred
    x-owner: schema-registry
    get:
      tags: [Event Catalog]
      operationId: listEventDefinitions
      summary: 列出事件定義
      description: 查詢平台已登錄的事件類型。這不是從 ClickHouse 查詢原始事件，而是查詢 PostgreSQL 控制面中的事件目錄。
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageCursor'
        - name: producer
          in: query
          description: 只回傳指定 Producer 的事件定義。
          schema:
            type: string
        - name: lifecycle_status
          in: query
          description: 事件生命週期狀態。
          schema:
            $ref: '#/components/schemas/LifecycleStatus'
      responses:
        '200':
          description: 事件定義清單。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDefinitionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Event Catalog]
      operationId: createEventDefinition
      summary: 建立事件定義
      description: 建立一個事件類型的治理資料。事件本身仍由業務服務透過 Outbox 與 Kafka 發布。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventDefinitionRequest'
      responses:
        '201':
          description: 建立成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDefinition'
        '409':
          description: 同一事件類型已存在，或版本衝突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/events/{eventType}:
    x-mvp-status: deferred
    x-owner: schema-registry
    parameters:
      - $ref: '#/components/parameters/EventType'
    get:
      tags: [Event Catalog]
      operationId: getEventDefinition
      summary: 取得事件定義
      description: 取得指定事件類型與其目前治理狀態。
      responses:
        '200':
          description: 事件定義。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDefinition'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/investigations/{investigationId}/evidence/events:
    x-mvp-status: phase-1.1
    x-owner: event-hunter
    parameters:
      - $ref: '#/components/parameters/InvestigationId'
    post:
      tags: [Investigation]
      operationId: attachInvestigationEvent
      summary: 將 Timeline event 加入案件
      description: |
        以 bounded `from`/`to` 到 ClickHouse 精確驗證 event 後，只將 EVENT reference 與
        checksum 附加到案件，不複製 event payload。同一案件重複附加同一 event 為
        idempotent no-op；新 evidence、related correlation 與 lock_version 在同一個
        PostgreSQL transaction 內完成。
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachInvestigationEventRequest'
      responses:
        '200':
          description: Event 已附加，或相同 reference 先前已存在。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachInvestigationEventResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
components:
  securitySchemes:
    demoSession:
      type: apiKey
      in: cookie
      name: eh_demo_session
      description: MVP 示範用的簽署 HttpOnly Cookie；不對應使用者資料表。
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 正式環境預留，由外部 OIDC／Identity Provider 核發；MVP 使用 `demoSession`。
    grafanaHmac:
      type: apiKey
      in: header
      name: X-Grafana-Alerting-Signature
      description: |
        Grafana Webhook Contact Point 產生的 hex HMAC-SHA256。伺服器必須搭配
        `X-Grafana-Alerting-Timestamp` 與 raw request body 驗證，使用 constant-time comparison。
  parameters:
    EventType:
      name: eventType
      in: path
      required: true
      description: 事件類型，例如 `PaymentCompleted`。
      schema:
        type: string
        minLength: 1
        maxLength: 200
        pattern: '^[A-Za-z][A-Za-z0-9._-]*$'
...
    InvestigationId:
      name: investigationId
      in: path
      required: true
      description: 調查案件 UUID。
      schema:
        type: string
        format: uuid
...
    PageSize:
      name: page_size
      in: query
      description: 每頁筆數；伺服器仍會套用最大上限。
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    PageCursor:
      name: page_cursor
      in: query
      description: 上一頁回應的 `next_cursor`，第一次查詢不傳。
      schema:
        type: string
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: GET 回應的 ETag，例如 `W/"investigation-123-lock-v7"`。版本不符時回傳 `409 Conflict`。
      schema:
        type: string
        minLength: 1
...
  headers:
    ETag:
      description: 資源版本，用於下一次更新的 `If-Match` 樂觀鎖。
      schema:
        type: string
        example: W/"investigation-123-lock-v7"
    RateLimitLimit:
      description: 目前固定窗口允許的最大請求數。
      schema:
        type: integer
    RateLimitRemaining:
      description: 目前固定窗口剩餘的請求數。
      schema:
        type: integer
    RateLimitReset:
      description: 距離目前窗口重設的秒數。
      schema:
        type: integer
    RetryAfter:
      description: 被限流後可重試前應等待的秒數。
      schema:
        type: integer
  responses:
    BadRequest:
      description: 請求格式或查詢條件錯誤。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request Schema 驗證失敗。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFound:
      description: 找不到指定資源。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 目前角色缺少執行此操作所需的權限。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 樂觀鎖版本衝突或目前狀態不允許此操作。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictError'
    RateLimited:
      description: 同一遠端位址在固定窗口內超過 API 請求上限。
      headers:
        RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    DependencyUnavailable:
      description: 必要的 PostgreSQL 或 ClickHouse 來源無法使用。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    DemoRole:
      type: string
      enum: [VIEWER, INVESTIGATOR, ADMIN]
...
    LifecycleStatus:
      type: string
      enum: [DRAFT, ACTIVE, DEPRECATED]
      description: 事件或 Schema 的生命週期。
...
    Severity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
      description: 案件或品質違規嚴重度。
...
    InvestigationStatus:
      type: string
      enum: [OPEN, INVESTIGATING, WAITING_APPROVAL, RESOLVED, CLOSED]
      description: 調查案件狀態；狀態轉移需由伺服器依狀態機驗證。
...
    Pagination:
      type: object
      required: [page_size]
      properties:
        page_size:
          type: integer
          minimum: 1
        next_cursor:
          type: [string, 'null']
          description: 下一頁游標；沒有下一頁時為 null。
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          description: 給使用者或工程師看的錯誤說明。
          example: 找不到指定的調查案件。
        request_id:
          type: string
          description: 供客服與日誌查找的請求 ID。
          example: req_01J5QZ7F9P
        details:
          type: object
          additionalProperties: true
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            field_errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string
    ConflictError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          required: [current_lock_version]
          properties:
            current_lock_version:
              type: integer
              format: int64
              description: 伺服器目前版本；客戶端應重新 GET 後再提交。
            current_resource:
              type: object
              additionalProperties: true
...
    EventDefinition:
      type: object
      required: [event_type, description, producer, owner, compatibility_policy, lifecycle_status, lock_version, created_at, updated_at]
      properties:
        event_type:
          type: string
          example: PaymentCompleted
        description:
          type: string
          example: 付款成功後由 Payment Service 發出的業務事件。
        producer:
          type: string
          example: payment-service
        consumers:
          type: array
          items:
            type: string
          example: [order-service, analytics-sink]
        owner:
          type: string
          example: payments-platform
        compatibility_policy:
          type: string
          enum: [BACKWARD, FORWARD, FULL, NONE]
          description: Schema 演進的相容性政策。
        lifecycle_status:
          $ref: '#/components/schemas/LifecycleStatus'
        current_schema_version:
          type: [integer, 'null']
          minimum: 1
        lock_version:
          type: integer
          format: int64
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEventDefinitionRequest:
      type: object
      required: [event_type, description, producer, owner, compatibility_policy]
      properties:
        event_type:
          type: string
          pattern: '^[A-Za-z][A-Za-z0-9._-]*$'
          example: PaymentCompleted
        description:
          type: string
          minLength: 1
        producer:
          type: string
          example: payment-service
        consumers:
          type: array
          items:
            type: string
        owner:
          type: string
        compatibility_policy:
          type: string
          enum: [BACKWARD, FORWARD, FULL, NONE]
        lifecycle_status:
          $ref: '#/components/schemas/LifecycleStatus'
          default: DRAFT
...
    PatternFinding:
      type: object
      additionalProperties: false
      required: [pattern_id, matched_conditions, severity, evidence_references, recommended_next_query]
      properties:
        finding_id:
          type: string
          format: uuid
          description: 已持久化 finding 的識別碼；同步分析中的暫時結果可能尚未提供。
        pattern_id:
          type: string
          example: payment-completed-without-shipment
        matched_conditions:
          type: array
          items:
            type: string
          example: [PAYMENT_COMPLETED_EXISTS, SHIPMENT_CREATED_MISSING_WITHIN_5M]
        severity:
          $ref: '#/components/schemas/Severity'
        evidence_references:
          type: array
          items:
            type: string
          example: [evt-001, evt-002, trace-500]
        recommended_next_query:
          type: string
          example: 查詢 shipping.events 是否在付款成功後建立 Shipment。
        query_template_id:
          type: string
          example: events.by_correlation.v1
        replay_conditions:
          type: array
          items:
            type: string
          description: 未來 Projection Rebuild／Sandbox Replay 可驗證的條件；MVP 只保存，不執行。
        feedback:
          $ref: '#/components/schemas/PatternFindingFeedback'
...
    PatternFindingFeedback:
      type: object
      additionalProperties: false
      required: [finding_id, status, actor_id, actor_role, updated_at, lock_version]
      properties:
        finding_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [UNREVIEWED, CONFIRMED, FALSE_POSITIVE, NEEDS_REVIEW]
        actor_id:
          type: string
          description: UNREVIEWED 時為空字串。
        actor_role:
          type: string
          description: UNREVIEWED 時為空字串。
        updated_at:
          type: [string, 'null']
          format: date-time
        lock_version:
          type: integer
          format: int64
          minimum: 0
...
    Evidence:
      type: object
      additionalProperties: false
      required: [id, evidence_type, reference, source, open_action, collected_at]
      properties:
        id:
          type: string
          format: uuid
        evidence_type:
          type: string
          enum: [EVENT, TRACE, LOG, METRIC, GRAFANA_ALERT, QUALITY_VIOLATION, PATTERN_FINDING, REPORT]
        reference:
          type: string
          description: 指向 ClickHouse、Trace、Log 或案件內部資料的穩定參照，不直接內嵌完整 PII。
        source:
          type: string
          enum: [CLICKHOUSE, TEMPO, LOKI, GRAFANA, PATTERN_ENGINE, REPORT_STORE, UNKNOWN]
          description: reference 的權威來源；Event Hunter 只保存參照與完整性資訊。
        open_action:
          type: string
          enum: [GRAFANA_EVENT, GRAFANA_TEMPO, GRAFANA_LOKI, GRAFANA_DASHBOARD, GRAFANA_ALERT, PATTERN_LIBRARY, NONE]
          description: 前端可依受信任 URL builder 執行的來源操作，不接受任意外部 URL。
        source_locator:
          type: string
          pattern: '^/alerting/(grafana/)?[A-Za-z0-9_-]+/(view|edit)$'
          description: 經後端驗證的相對來源路徑；不得包含 origin、credentials 或任意外部 URL。
        source_org_id:
          type: integer
          minimum: 1
          description: 來源 Grafana organization ID；只與受信任的 relative source_locator 組合。
        collected_at:
          type: string
          format: date-time
        checksum:
          type: [string, 'null']
          description: 證據內容雜湊，用於確認報告產生後資料未被替換。
...
    InvestigationCase:
      type: object
      required: [id, case_no, title, severity, status, allowed_transitions, correlation_id, incident_from, incident_to, incident_window_source, priority, tags, related_correlation_ids, last_updated_by, sla_status, sla_due_at, collaboration_notes, evidence, pattern_findings, lock_version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        case_no:
          type: string
          example: INC-2026-0001
        title:
          type: string
          example: 訂單付款成功但未建立出貨單
        severity:
          $ref: '#/components/schemas/Severity'
        status:
          $ref: '#/components/schemas/InvestigationStatus'
        allowed_transitions:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/InvestigationStatus'
          description: |
            目前案件可執行的目標狀態，由 aggregate 產生。`CLOSED` 表示可呼叫專用
            closeInvestigation operation；不得以 updateInvestigation 直接設為 CLOSED。
        correlation_id:
          type: string
          example: ORDER-1001
        incident_from:
          type: string
          format: date-time
          description: 案件建立時保存的不可變基準窗口起點。
        incident_to:
          type: string
          format: date-time
          description: 案件建立時保存的不可變基準窗口終點；與 incident_from 的距離最多七天。
        incident_window_source:
          type: string
          enum: [TIMELINE_SEARCH, MANUAL_DEFAULT, GRAFANA_ALERT, LEGACY_CREATED_AT]
          description: 基準窗口的可追溯推導來源。
        assignee:
          type: [string, 'null']
          description: Phase 1.1 的 owner；沿用 assignee 欄位名稱避免重複責任欄位。
          example: oncall-payments
        priority:
          $ref: '#/components/schemas/CasePriority'
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 50
        related_correlation_ids:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 200
        last_updated_by:
          type: string
        sla_status:
          $ref: '#/components/schemas/CaseSLAStatus'
        sla_due_at:
          type: string
          format: date-time
        root_cause:
          type: [string, 'null']
        resolution_summary:
          type: [string, 'null']
        fixed_version:
          type: [string, 'null']
        notes:
          type: [string, 'null']
          readOnly: true
          deprecated: true
          description: Phase 1 舊資料的唯讀相容欄位；新筆記一律使用 collaboration_notes。
        collaboration_notes:
          type: array
          items:
            $ref: '#/components/schemas/InvestigationNote'
        workflow_id:
          type: [string, 'null']
          description: Temporal Workflow ID；不保存 Temporal Server 內部資料作為本平台資料表。
        pattern_findings:
          type: array
          items:
            $ref: '#/components/schemas/PatternFinding'
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/Evidence'
        lock_version:
          type: integer
          format: int64
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: [string, 'null']
          format: date-time
...
    CasePriority:
      type: string
      enum: [P0, P1, P2, P3]
      description: P0/P1/P2/P3 的目標處理時間分別為 1/4/24/72 小時。
    CaseSLAStatus:
      type: string
      enum: [ON_TRACK, DUE_SOON, BREACHED, COMPLETED]
    InvestigationNote:
      type: object
      additionalProperties: false
      required: [id, body, author_id, author_role, created_at]
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
          minLength: 1
          maxLength: 2000
        author_id:
          type: string
        author_role:
          $ref: '#/components/schemas/DemoRole'
        created_at:
          type: string
          format: date-time
...
    AttachInvestigationEventRequest:
      type: object
      additionalProperties: false
      required: [event_id, from, to]
      properties:
        event_id:
          type: string
          minLength: 1
          maxLength: 200
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
          description: 必須晚於 from，且查詢時間窗不得超過七天。
    AttachInvestigationEventResponse:
      type: object
      additionalProperties: false
      required: [investigation, evidence, attached]
      properties:
        investigation:
          $ref: '#/components/schemas/InvestigationCase'
        evidence:
          $ref: '#/components/schemas/Evidence'
        attached:
          type: boolean
          description: true 表示本次新增 reference；false 表示相同 evidence 已存在。
...
    EventDefinitionPage:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EventDefinition'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
security:
  - demoSession: []
  - bearerAuth: []